+++ udata_gouvfr/theme/__init__.py
@@ -45,7 +45,10 @@ gouvfr_menu = nav.Bar('gouvfr_menu', [
     nav.Item(_('Reuses'), 'reuses.list'),
     nav.Item(_('Organizations'), 'organizations.list'),
     nav.Item(_('Dashboard'), 'site.dashboard'),
-    nav.Item(_('Documentation'), None, url='https://doc.data.gouv.fr'),
+    nav.Item(_('Documentation'), None, url='https://doc.data.gouv.fr', items=[
+        nav.Item(_('Technical documentation'), None, url='https://doc.data.gouv.fr'),
+        nav.Item(_('Open data guides'), None, url='https://guides.etalab.gouv.fr'),
+    ]),
     nav.Item(_('News'), 'posts.list'),
 ])